+++ src/filters/transform/MpaDecFilter/MpaDecFilter.cpp
@@ -950,12 +950,12 @@ HRESULT CMpaDecFilter::ProcessAC3()
 		if (m_DolbyDigitalMode != DD_TRUEHD && m_DolbyDigitalMode != DD_MLP && (*((__int16*)p) == 0x770b)) { /* AC3-EAC3 syncword */
 			BYTE bsid = p[5] >> 3;
 			if (bsid <= 10) {
-				if (FAILED (hr = ProcessA52 (p, end-p, size, fEnoughData))) {
+				if (FAILED (hr = ProcessA52 (p, int(end-p), size, fEnoughData))) {
 					return hr;
 				}
 				m_DolbyDigitalMode = DD_AC3;
 			} else if (bsid <= 16) {
-				DeliverFFmpeg(CODEC_ID_EAC3, p, end-p, size);
+				DeliverFFmpeg(CODEC_ID_EAC3, p, int(end-p), size);
 				if (size > 0) {
 					m_DolbyDigitalMode = DD_EAC3;
 				}
@@ -970,9 +970,9 @@ HRESULT CMpaDecFilter::ProcessAC3()
 			m_DolbyDigitalMode = DD_TRUEHD;
 
 			if (nLenght >= 4) {
-				DeliverFFmpeg(CODEC_ID_TRUEHD, p, end-p, size);
+				DeliverFFmpeg(CODEC_ID_TRUEHD, p, int(end-p), size);
 				if (size<0) {
-					size = end-p;
+					size = (int)(end-p);
 				}
 			}
 		} else if ( (*((__int32*)(p+4)) == 0xbb6f72f8) || // MLP major sync frame
@@ -982,9 +982,9 @@ HRESULT CMpaDecFilter::ProcessAC3()
 			m_DolbyDigitalMode = DD_MLP;
 
 			if (nLenght >= 4) {
-				DeliverFFmpeg(CODEC_ID_MLP, p, end-p, size);
+				DeliverFFmpeg(CODEC_ID_MLP, p, int(end-p), size);
 				if (size<0) {
-					size = end-p;
+					size = (int)(end-p);
 				}
 			}
 		} else {
@@ -1023,7 +1023,7 @@ HRESULT CMpaDecFilter::ProcessFFmpeg(enum CodecID nCodecId)
 	BYTE* end = p + m_buff.GetCount();
 
 	int size = 0;
-	hr = DeliverFFmpeg(nCodecId, p, end-p, size);
+	hr = DeliverFFmpeg(nCodecId, p, int(end-p), size);
 	if (FAILED(hr)) {
 		if (!(nCodecId == CODEC_ID_AAC || nCodecId == CODEC_ID_AAC_LATM)) {
 			m_buff.RemoveAll();
